fix: address CodeQL findings in VisualMissionItem, QGCCachedTileSet, SubtitleWriter
VisualMissionItem: removed an unused copy constructor that assigned into a
not-yet-fully-constructed object (*this = other before _commonInit()).
If that assignment path ever called a virtual like setDirty(), it would
dispatch on an incomplete vtable - undefined behavior. No callers existed
for this constructor.

QGCCachedTileSet: three tile-size estimates multiplied 32-bit operands and
assigned the (potentially already-truncated) result into a quint64,
overflowing before the destination's width could help. Cast the narrower
operand to quint64 before multiplying in each case. Added a regression
test exercising a >4GiB estimate.

SubtitleWriter: nRows was declared float for no reason and fed a
floating-point ceil() division; widened the row-count division's result
to qsizetype (used downstream in a multiplication) and replaced the
float/ceil() combination with exact integer ceiling division.

The remaining 13 of 18 findings in this scan pass (type confusion in
StandardModes/RequestMessageState/GimbalController/QGCCameraManager,
narrow/wide loop comparisons in VehicleCameraControl/OnboardLogController,
and allocation-error-handling patterns in PX4AutoPilotPlugin/
AirframeComponentController/MAVLinkLogManager/FirmwareUpgradeController)
were investigated and are false positives - see PR description for the
per-finding reasoning.

Verified: ninja build clean; full ctest -L Unit (328/328); pre-commit
clean on all changed files.

+++ src/QtLocationPlugin/QGCCachedTileSet.cpp
@@ -117,7 +117,7 @@ void QGCCachedTileSet::_doneWithDownload()
             qCWarning(QGCCachedTileSetLog) << "_savedTileCount=0";
         }
 
-        setUniqueTileSize(_uniqueTileCount * avg);
+        setUniqueTileSize(static_cast<quint64>(_uniqueTileCount) * avg);
     }
 
     setDownloading(false);
@@ -243,8 +243,8 @@ void QGCCachedTileSet::_networkReplyFinished()
 
     if (_savedTileCount % 10 == 0) {
         const quint32 avg = _savedTileSize / _savedTileCount;
-        setTotalTileSize(avg * _totalTileCount);
-        setUniqueTileSize(avg * _uniqueTileCount);
+        setTotalTileSize(static_cast<quint64>(avg) * _totalTileCount);
+        setUniqueTileSize(static_cast<quint64>(avg) * _uniqueTileCount);
     }
 
     _prepareDownload();
